# contracts/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MautoDesk API",
    "description": "Multi-tenant dealership management platform.\n\n**Authentication.** Bearer access tokens. Tenant scope is carried in the token's `tenant` claim; supplying a tenant identifier in a header, query string, or body has no effect and is never honoured.\n\n**Errors.** All non-2xx responses are RFC 9457 `application/problem+json`. A resource belonging to another tenant is reported as `404 Not Found` — the API does not confirm the existence of records outside the caller's tenant.\n\n**Money** is transported as a decimal string (`\"28995.00\"`), never a JSON number, so no client can round it through a floating-point value.",
    "license": {
      "name": "Proprietary",
      "url": "https://mautodesk.com/legal/api-terms"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.mautodesk.com",
      "description": "Production"
    },
    {
      "url": "https://staging-api.mautodesk.com",
      "description": "Staging"
    },
    {
      "url": "http://localhost:5080",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Authenticate with email and password",
        "description": "Returns tokens only when no second factor is outstanding. Because MFA is mandatory, the normal responses are mfaRequired or mfaEnrolmentRequired. The response for an unknown address is identical to the response for a wrong password, and the timing is equalised, so this endpoint cannot be used to discover which addresses have accounts.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, or invalid credentials.",
            "content": {
              "application/problem+json": { }
            }
          }
        }
      }
    },
    "/api/v1/auth/mfa/verify": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Complete authentication with a TOTP code",
        "operationId": "verifyMfa",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyMfaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, or invalid credentials.",
            "content": {
              "application/problem+json": { }
            }
          }
        }
      }
    },
    "/api/v1/auth/mfa/enrol": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Confirm a new authenticator and finish signing in",
        "operationId": "confirmMfaEnrolment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyMfaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, or invalid credentials.",
            "content": {
              "application/problem+json": { }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Exchange a refresh token for a new pair",
        "description": "Refresh tokens rotate on every use. Presenting a token that has already been rotated is treated as evidence of theft: the entire token family is revoked and a fresh sign-in is required.",
        "operationId": "refreshToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair2"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, or invalid credentials.",
            "content": {
              "application/problem+json": { }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Revoke the current session",
        "operationId": "logout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Missing, expired, or invalid credentials.",
            "content": {
              "application/problem+json": { }
            }
          },
          "403": {
            "description": "The principal lacks the required permission.",
            "content": {
              "application/problem+json": { }
            }
          }
        }
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "summary": "The authenticated principal, tenant, and effective permissions",
        "operationId": "getCurrentUser",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUserDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired, or invalid credentials.",
            "content": {
              "application/problem+json": { }
            }
          },
          "403": {
            "description": "The principal lacks the required permission.",
            "content": {
              "application/problem+json": { }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthResponse": {
        "required": [
          "outcome",
          "tokens",
          "challengeToken",
          "enrolmentSecret",
          "enrolmentUri"
        ],
        "type": "object",
        "properties": {
          "outcome": {
            "type": "string"
          },
          "tokens": {
            "$ref": "#/components/schemas/TokenPair"
          },
          "challengeToken": {
            "type": "string",
            "nullable": true
          },
          "enrolmentSecret": {
            "type": "string",
            "nullable": true
          },
          "enrolmentUri": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "CurrentUserDto": {
        "required": [
          "id",
          "email",
          "firstName",
          "lastName",
          "displayName",
          "tenant",
          "roles",
          "permissions",
          "mfaEnrolled"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "tenant": {
            "$ref": "#/components/schemas/TenantDto"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "mfaEnrolled": {
            "type": "boolean"
          }
        }
      },
      "LoginRequest": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "RefreshRequest": {
        "required": [
          "refreshToken"
        ],
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "TenantDto": {
        "required": [
          "id",
          "slug",
          "legalName",
          "stateCode"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string"
          },
          "legalName": {
            "type": "string"
          },
          "stateCode": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "TokenPair": {
        "required": [
          "accessToken",
          "refreshToken",
          "tokenType",
          "expiresIn"
        ],
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "tokenType": {
            "type": "string"
          },
          "expiresIn": {
            "type": "integer",
            "format": "int32"
          }
        },
        "nullable": true
      },
      "TokenPair2": {
        "required": [
          "accessToken",
          "refreshToken",
          "tokenType",
          "expiresIn"
        ],
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "tokenType": {
            "type": "string"
          },
          "expiresIn": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "VerifyMfaRequest": {
        "required": [
          "challengeToken",
          "code"
        ],
        "type": "object",
        "properties": {
          "challengeToken": {
            "type": "string",
            "nullable": true
          },
          "code": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Access token from /auth/login. The tenant is carried in the token's `tenant` claim and cannot be overridden by any request parameter.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": [ ]
    }
  ],
...
}